requirements.txt
zenml[server]==0.84.0
requests
beautifulsoup4
python-dotenv
pymongo
qdrant-client
streamlit
openai
rapidfuzz
sentence-transformers
jinja2
transformers
datasets
peft
accelerate
bitsandbytes
